fix(review): PR-B — gmnspy CLI cleanup (F3, F7, F9)
Batch A review surfaced three issues in packages/gmnspy/gmnspy/cli/app.py:

F3 — _ScopeErrors context manager caught by exc_type.__name__ string-match,
silently swallowed every stdlib ValueError, and always told users to install
gmnspy[clean] even for non-clean imports. DELETED. Each scope/clean command
now catches its real domain exception class explicitly (ScopeError /
CleanError, reached through the dynamically-imported module so the
import-linter contract gmnspy.cli ↛ gmnspy.{clean,scope} still holds).
Bare ValueError no longer swallowed.

F7 — app.py was ~950 LOC with 9 sub-apps as nested closures inside a
521-line factory. Split per sub-app: each command lives in
packages/gmnspy/gmnspy/cli/commands/<name>.py and exposes a register(app)
hook; cli/app.py is now a ~65-line orchestrator. Readers hunting for
'gmnspy clean simplify-geometry' open commands/clean.py directly instead
of scroll-searching a giant file.

F9 — importlib.import_module() was repeated 5 times with 3 different error
shapes (server, mcp, clean, scope, indexes). Consolidated into
gmnspy.cli._extras.require_extra() — one helper, one consistent install hint
on stderr with the correct extra name, one place to add caching / dev-time
warnings later.

Tests: +3 in test_cli_extras.py pinning require_extra's contract (returns
present module; exits 1 with red 'pip install gmnspy[<extra>]' hint on
missing module; original ImportError text included for debuggability).
All existing CLI tests pass unchanged — the typer surface is identical,
only the internal file layout changed.

Verification:
- uv run pytest                       → 879 passed, 48 skipped (+3 vs baseline)
- uv run ruff check/format            → clean
- uv run lint-imports                 → 2/2 contracts kept
- uv run python scripts/lint_no_sql.py → clean
- Smoke: gmnspy --help / info / clean simplify-geometry --dry-run /
  scope from-nodes / doctor all OK
